Move insets to Redwood-only module
