+++ redwood-composeui/build.gradle
@@ -14,6 +14,7 @@ kotlin {
   sourceSets {
     commonMain {
       dependencies {
+        implementation libs.jetbrains.compose.foundation
         api projects.redwoodCompose
         implementation projects.redwoodWidgetCompose
       }
