+++ redwood-composeui/src/iosMain/kotlin/app/cash/redwood/composeui/insets.kt
@@ -13,11 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.treehouse.composeui
+package app.cash.redwood.composeui
 
 import androidx.compose.runtime.Composable
+import app.cash.redwood.ui.Default
+import app.cash.redwood.ui.Density
 import app.cash.redwood.ui.Margin
+import kotlinx.cinterop.useContents
+import platform.UIKit.UIApplication
 
-/** Return the device's insets in screen density-independent pixels. */
 @Composable
-internal expect fun safeAreaInsets(): Margin
+public actual fun safeAreaInsets(): Margin {
+  val keyWindow = UIApplication.sharedApplication.keyWindow ?: return Margin.Zero
+  return keyWindow.safeAreaInsets.useContents {
+    with(Density.Default) {
+      Margin(left.toDp(), right.toDp(), top.toDp(), bottom.toDp())
+    }
+  }
+}
