+++ redwood-treehouse-host-composeui/src/commonMain/kotlin/app/cash/redwood/treehouse/composeui/TreehouseContent.kt
@@ -27,6 +27,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalDensity
+import app.cash.redwood.composeui.safeAreaInsets
 import app.cash.redwood.treehouse.AppService
 import app.cash.redwood.treehouse.CodeListener
 import app.cash.redwood.treehouse.StateSnapshot
